Refactor QuizlyApp into its quizly_app folder inside presentation folder
