+++ test/widget_test.dart
@@ -5,10 +5,8 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:quizly/main.dart';
+import 'package:quizly/presentation/quizly_app/quizly_app.dart';
 
 void main() {
   testWidgets('Quizly starter smoke test', (WidgetTester tester) async {
